pyproject.toml
[project]
name = "simplified-ai-parser"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4==4.12.3",
    "fastapi==0.111.0",
    "lxml>=5.1.0",
    "mammoth==1.8.0",
    "markdownify==0.14.1",
    "openpyxl>=3.1.5",
    "pillow>=11.1.0",
    "pydantic-settings>=2.6.1",
    "python-pptx>=0.6.23",
    "python-multipart>=0.0.19",
    "uvicorn==0.30.1",
]
